Got the basic experiment system working.
Added a `main` function so we can call this from the command line with `lein`, which makes it easier to collect the printed results. I also added some labelling using metadata and some symbol resolution craziness; it's not super pretty, but it works. It definitely makes it clear that I should have included the names of problems in the problem instance maps – sigh.